% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter-estimation.R
\name{augment_data}
\alias{augment_data}
\title{Augment data frame for missing counterfactuals}
\usage{
augment_data(data1, missingind, ID.var, formula.predict, Nout)
}
\arguments{
\item{data1}{Data frame in question}

\item{missingind}{vector the same number of rows as data1, indicator for the missingness}

\item{ID.var}{character vector specifying the ID variable in data1}

\item{formula.predict}{Formula specifying the model}

\item{Nout}{integer, the number of imputed observations per missing one}
}
\value{
A list containing two dataframes, the unaugmented data, and the augmented data
}
\description{
Impute missing counterfactuals using a linear model on the
baseline measurements
}
