% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation.R
\name{cumincfun}
\alias{cumincfun}
\title{Compute the cumulative incidence}
\usage{
cumincfun(estbetas, timet, datain)
}
\arguments{
\item{estbetas}{Fitted flexsurv object as returned by \link{estimate_flexparams}}

\item{timet}{The time at which the cumulative incidence is to be computed}

\item{datain}{data frame containing all variables in the model}
}
\description{
Compute the cumulative incidence
}
