% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation.R
\name{gcompute_cuminc}
\alias{gcompute_cuminc}
\title{G compute the cumulative incidence function}
\usage{
gcompute_cuminc(dataou1, fittedmod, timet, treatment.var, surrogate.var,
  nn = 100)
}
\arguments{
\item{dataou1}{Augmented data as returned by \link{augment_data}}

\item{fittedmod}{Parameter estimates as returned by \link{estimate_flexparams}}

\item{timet}{numeric time at which cumulative incidence is computed}

\item{treatment.var}{string denoting the name of the treatment variable}

\item{surrogate.var}{string denoting the name of the surrogate variable}

\item{nn}{number of values to sample for each observed S.1}
}
\value{
A data frame with the variables that go into the model, the estimated cumulative
incidence and an indicator for the group to average over
}
\description{
assuming S(1), W are jointly normal
sample from W | S(1), estimate CI, then average over the Ws
at values of S(1) that are observed in the vaccine arm
}
